Factor Digest-related traits out of 'digest' module
...allowing us to (once again) `pub extern crate digest`

This flattens out the crate's API, adding a `Digest*` prefix to all
traits that deal with `Digest` types, and re-exporting them from the
crate root.
