--- signature-crate/src/digest/verifier.rs
@@ -1,24 +1,0 @@
-//! Support for verifying messages which have been prehashed messages using
-//! the `Digest` trait.
-//!
-//! For use signature algorithms that support an Initialize-Update-Finalize
-//! (IUF) API, such as ECDSA or Ed25519ph.
-
-use super::Digest;
-use crate::{error::Error, signature::Signature};
-use digest::generic_array::GenericArray;
-
-/// Verify the provided signature for the given prehashed message `Digest`
-/// is authentic.
-pub trait Verifier<D, S>
-where
-    D: Digest,
-    S: Signature,
-{
-    /// Verify the signature against the given `Digest`
-    fn verify_digest(
-        &self,
-        digest: GenericArray<u8, D::OutputSize>,
-        signature: &S,
-    ) -> Result<(), Error>;
-}
